Add documentation and move checkpointer into file
