+++ src/adios4dolfinx/__init__.py
@@ -12,6 +12,8 @@ from .checkpointing import (
     read_mesh,
     write_function,
     write_mesh,
+    write_meshtags,
+    read_meshtags,
     snapshot_checkpoint,
 )
 from .legacy_readers import (
@@ -27,6 +29,8 @@ __email__ = meta["Author-email"]
 __program_name__ = meta["Name"]
 
 __all__ = [
+    "write_meshtags",
+    "read_meshtags",
     "read_mesh",
     "write_mesh",
     "read_function_from_legacy_h5",
